refactor : sections로 로직 이동
